+++ lib/src/ast/value/table.dart
@@ -10,36 +10,42 @@ import '../visitor/value.dart';
 
 /// AST node that represents a TOML inline table.
 ///
-///     inline-table =
-///         inline-table-open [ inline-table-keyvals ] inline-table-close
+///     inline-table = inline-table-open [ inline-table-keyvals ]
+///                    ws-comment-newline inline-table-close
 ///
-///     inline-table-keyvals = keyval [ inline-table-sep inline-table-keyvals ]
+///     inline-table-keyvals =  ws-comment-newline keyval ws-comment-newline
+///                             inline-table-sep inline-table-keyvals
+///     inline-table-keyvals =/ ws-comment-newline keyval ws-comment-newline
+///                             [ inline-table-sep ]
 @immutable
 class TomlInlineTable extends TomlValue {
   /// The opening delimiter of inline tables.
   ///
-  ///     inline-table-open  = %x7B ws     ; {
+  ///     inline-table-open  = %x7B  ; {
   static final String openingDelimiter = '{';
 
   /// The separator for the key/value pairs in inline tables.
   ///
-  ///     inline-table-sep   = ws %x2C ws  ; , Comma
+  ///     inline-table-sep   = %x2C  ; , Comma
   static final String separator = ',';
 
   /// The closing delimiter of inline tables.
   ///
-  ///     inline-table-close = ws %x7D     ; }
+  ///     inline-table-close = %x7D  ; }
   static final String closingDelimiter = '}';
 
   /// Parser for a TOML inline-table.
-  ///
-  /// Trailing commas are currently not allowed.
-  /// See https://github.com/toml-lang/toml/pull/235#issuecomment-73578529
   static final Parser<TomlInlineTable> parser = TomlKeyValuePair.parser
-      .starSeparated(tomlWhitespace & char(separator) & tomlWhitespace)
-      .trim(tomlWhitespaceChar)
-      .skip(before: char(openingDelimiter), after: char(closingDelimiter))
+      .skip(
+        before: tomlWhitespaceCommentNewline,
+        after: tomlWhitespaceCommentNewline,
+      )
+      .plusSeparated(char(separator))
+      .skip(after: char(separator).optional())
       .map(discardSeparators)
+      .optionalWith(<TomlKeyValuePair>[])
+      .skip(after: tomlWhitespaceCommentNewline)
+      .skip(before: char(openingDelimiter), after: char(closingDelimiter))
       .map(TomlInlineTable.new);
 
   /// The key/value pairs of the inline table.